Normalize line endings and restore WordPress files
Restored complete WordPress installation with normalized CRLF line endings for Windows compatibility. Includes:
- WordPress core files (wp-admin, wp-includes, wp-content)
- Jannah theme with all demo organization work
- Saved school demo files in permanent location
- All uploaded demo images and content

+++ wp-content/themes/jannah/framework/admin/classes/class-tielabs-notices.php
@@ -295,7 +295,7 @@ if( ! class_exists( 'TIELABS_NOTICES' ) ){
 				}
 				else{
 
-					if( ( time() - get_option( $install_date ) ) < ( 5 * DAY_IN_SECONDS ) ){
+					if( ( time() - (int)get_option( $install_date ) ) < ( 5 * DAY_IN_SECONDS ) ){
 						return false;
 					}
